+++ NEWS.md
@@ -1,12 +1,22 @@
 # ExPanDaR 0.2.0.9000
 
+Extensions:
+
+* Removed the requirement for at least one non-numerical variable
+
 * Added by group violin plot
 
 * Included '!' and 'is.na()' as allowed functions for user defined variables
 
+
+Bug fixes:
+
 * Fixed a sorting bug in long variable definition construction
 
 * Fixed a bug in ExPanD() extreme obs plot (was sorting on grouping variable)
+
+
+Minor issues:
 
 * Removed the dependency with the CodeDepends package